# workflow/openapi.yaml
...
openapi: 3.1.0
info:
  title: GW/AP Debug Platform Workflow API
  version: 0.9.0
servers:
- url: http://localhost:8000/api/v1
...
paths:
...
  /jobs/{job_id}:
    get:
      summary: Get Job
      operationId: getBackgroundJob
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          description: Role, ownership, or resource scope does not authorize this operation
      x-allowed-roles:
      - ADMIN
      - EXPERT
      - ENGINEER
      - VIEWER
      x-agent-callable: true
...
  /cases/{case_id}/agent-runs/{run_id}:
    get:
      tags:
      - diagnostics
      summary: Get Case Agent Run
      operationId: getCaseAgentRun
      parameters:
      - name: case_id
        in: path
        required: true
        schema:
          type: string
          title: Case Id
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Case Agent Run Api V1 Cases  Case Id  Agent Runs  Run Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          description: Role, ownership, or resource scope does not authorize this operation
      x-allowed-roles:
      - ADMIN
      - EXPERT
      - ENGINEER
      - VIEWER
      x-agent-callable: true
...
  /agent-runs:
    get:
      tags:
      - agent-runs
      summary: List Agent Runs
      operationId: listAgentRuns
      parameters:
      - name: case_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Case Id
      - name: resource_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Resource Id
      - name: operation
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Operation
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 32
          - type: 'null'
          title: Status
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response List Agent Runs Api V1 Agent Runs Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          description: Role, ownership, or resource scope does not authorize this operation
      x-allowed-roles:
      - ADMIN
      - EXPERT
      x-agent-callable: true
      description: ADMIN/EXPERT execution-trace operation. Expose only content-safe metadata or a redacted replay
        payload; never publish knowledge or reveal model credentials.
  /agent-runs/{run_id}:
    get:
      tags:
      - agent-runs
      summary: Get Agent Run
      operationId: getAgentRun
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Agent Run Api V1 Agent Runs  Run Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          description: Role, ownership, or resource scope does not authorize this operation
      x-allowed-roles:
      - ADMIN
      - EXPERT
      x-agent-callable: true
      description: ADMIN/EXPERT execution-trace operation. Expose only content-safe metadata or a redacted replay
        payload; never publish knowledge or reveal model credentials.
  /agent-runs/{run_id}/replay:
    post:
      tags:
      - agent-runs
      summary: Replay Agent Run
      operationId: replayAgentRun
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Replay Agent Run Api V1 Agent Runs  Run Id  Replay Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          description: Role, ownership, or resource scope does not authorize this operation
      x-allowed-roles:
      - ADMIN
      - EXPERT
      x-agent-callable: true
      description: ADMIN/EXPERT execution-trace operation. Expose only content-safe metadata or a redacted replay
        payload; never publish knowledge or reveal model credentials.
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    JobOut:
      properties:
        id:
          type: string
          title: Id
        kind:
          type: string
          title: Kind
        status:
          type: string
          title: Status
        progress:
          type: integer
          title: Progress
        message:
          type: string
          title: Message
        result_json:
          type: string
          title: Result Json
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        idempotency_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency Key
        attempt:
          type: integer
          title: Attempt
        max_attempts:
          type: integer
          title: Max Attempts
        available_at:
          type: string
          format: date-time
          title: Available At
        lease_owner:
          anyOf:
          - type: string
          - type: 'null'
          title: Lease Owner
        lease_expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Lease Expires At
        heartbeat_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Heartbeat At
        deadline_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deadline At
        timeout_seconds:
          type: integer
          title: Timeout Seconds
        resource_limits_json:
          type: string
          title: Resource Limits Json
        dead_letter_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Dead Letter At
        dead_letter_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Dead Letter Reason
        created_at:
          type: string
          format: date-time
          title: Created At
        started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Started At
        completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Completed At
      type: object
      required:
      - id
      - kind
      - status
      - progress
      - message
      - result_json
      - error_message
      - idempotency_key
      - attempt
      - max_attempts
      - available_at
      - lease_owner
      - lease_expires_at
      - heartbeat_at
      - deadline_at
      - timeout_seconds
      - resource_limits_json
      - dead_letter_at
      - dead_letter_reason
      - created_at
      - started_at
      - completed_at
      title: JobOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
security:
- ApiKey: []